major cleanup of all the code

+++ src/stream_buffer.h
@@ -1,17 +1,23 @@
+/*
+ * File:  stream_buffer.h
+ * Author:  mikolas
+ * Created on:  Sat Jul 29 18:49:31 CEST 2023
+ * Copyright (C) 2023, Mikolas Janota
+ */
 #pragma once
+
+/* this is adapted from minisat's parser */
 
 #define CHUNK_LIMIT 1048576
 
 #include <cstdio>
-#include <sstream>
-#include <string>
 #include <zlib.h>
 class StreamBuffer {
   protected:
     gzFile in;
     char *buf;
-    int pos;
-    int size;
+    size_t pos;
+    size_t size;
 
     void assureLookahead() {
         if (pos >= size) {
